# site/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Business Performance Analyst Agent",
...
  "version": "1.0"
 },
 "paths": {
...
  "/api/run": {
   "get": {
    "tags": [
     "Agent runs"
    ],
    "summary": "Ask the agent a question (streamed)",
    "description": "The agent plans, runs read-only tools, and answers. `done.result` holds `answer`, `queries` (the exact\nSQL or code, from the tool log), `charts` (PNG, base64), `tokens`, `seconds` and `model`. Follow-up\nquestions remember the last answers until `POST /api/new-chat`.",
    "operationId": "run_api_run_get",
    "parameters": [
     {
      "name": "mode",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "`live` (the AI), `replay` (a recorded live run) or `dry` (no AI, scripted).",
       "enum": [
        "live",
        "replay",
        "dry"
       ],
       "example": "replay",
       "default": "dry",
       "title": "Mode"
      },
      "description": "`live` (the AI), `replay` (a recorded live run) or `dry` (no AI, scripted)."
     },
     {
      "name": "case",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "A demo track's id (from `/api/meta` `tracks`), e.g. `northwest_margin`. Takes priority over `question`.",
       "default": "",
       "title": "Case"
      },
      "description": "A demo track's id (from `/api/meta` `tracks`), e.g. `northwest_margin`. Takes priority over `question`."
     },
     {
      "name": "question",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Your own question, in plain English. `live` mode only, unless it matches a demo track.",
       "examples": [
        "What were total sales by territory last year?"
       ],
       "default": "",
       "title": "Question"
      },
      "description": "Your own question, in plain English. `live` mode only, unless it matches a demo track."
     },
     {
      "name": "delay",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "description": "Seconds to pause after each step, so a replay or dry run is easy to follow (used by the video recorder).",
       "default": 0.0,
       "title": "Delay"
      },
      "description": "Seconds to pause after each step, so a replay or dry run is easy to follow (used by the video recorder)."
     }
    ],
    "responses": {
     "200": {
      "description": "A stream of Server-Sent Events: see **Streaming runs** above.",
      "content": {
       "text/event-stream": {
        "example": "data: {\"type\": \"start\", \"question\": \"...\", \"mode\": \"replay\"}\n\ndata: {\"type\": \"step\", \"n\": 1, \"tool\": \"run_sql\", \"summary\": \"10 rows back, saved as q1\", \"body\": \"SELECT ...\"}\n\ndata: {\"type\": \"done\", \"result\": {\"answer\": \"...\", \"queries\": [...], \"charts\": [...]}}\n\n"
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/briefing": {
   "get": {
    "tags": [
     "Agent runs"
    ],
    "summary": "Write the weekly leadership briefing (streamed)",
    "description": "A `kpis` event comes first: the KPI table straight from SQL, final before the AI starts. `done.result`\nholds `headline` and the sections in `parts`, `kpis`, `charts`, `queries`, and `html`: the whole page,\nready to save or email.",
    "operationId": "run_briefing_api_briefing_get",
    "parameters": [
     {
      "name": "mode",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "`live` (the AI), `replay` (a recorded live run) or `dry` (no AI, scripted).",
       "enum": [
        "live",
        "replay",
        "dry"
       ],
       "example": "replay",
       "default": "dry",
       "title": "Mode"
      },
      "description": "`live` (the AI), `replay` (a recorded live run) or `dry` (no AI, scripted)."
     },
     {
      "name": "week",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Monday the week starts, `YYYY-MM-DD` (from `/api/meta` `weeks`). Default: the latest full week.",
       "examples": [
        "2025-06-23"
       ],
       "default": "",
       "title": "Week"
      },
      "description": "Monday the week starts, `YYYY-MM-DD` (from `/api/meta` `weeks`). Default: the latest full week."
     },
     {
      "name": "delay",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "description": "Seconds to pause after each step, so a replay or dry run is easy to follow (used by the video recorder).",
       "default": 0.0,
       "title": "Delay"
      },
      "description": "Seconds to pause after each step, so a replay or dry run is easy to follow (used by the video recorder)."
     }
    ],
    "responses": {
     "200": {
      "description": "A stream of Server-Sent Events: see **Streaming runs** above.",
      "content": {
       "text/event-stream": {
        "example": "data: {\"type\": \"start\", \"question\": \"...\", \"mode\": \"replay\"}\n\ndata: {\"type\": \"step\", \"n\": 1, \"tool\": \"run_sql\", \"summary\": \"10 rows back, saved as q1\", \"body\": \"SELECT ...\"}\n\ndata: {\"type\": \"done\", \"result\": {\"answer\": \"...\", \"queries\": [...], \"charts\": [...]}}\n\n"
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/scorecard": {
   "get": {
    "tags": [
     "Agent runs"
    ],
    "summary": "Mark the agent's answers (streamed)",
    "description": "Asks the brief's four questions and three guardrail traps, and streams one `result` event per question,\nmarked against answers worked out from the database.",
    "operationId": "scorecard_api_scorecard_get",
    "parameters": [
     {
      "name": "mode",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "`live` (the AI), `replay` (a recorded live run) or `dry` (no AI, scripted).",
       "enum": [
        "live",
        "replay",
        "dry"
       ],
       "example": "replay",
       "default": "dry",
       "title": "Mode"
      },
      "description": "`live` (the AI), `replay` (a recorded live run) or `dry` (no AI, scripted)."
     }
    ],
    "responses": {
     "200": {
      "description": "A stream of Server-Sent Events: see **Streaming runs** above.",
      "content": {
       "text/event-stream": {
        "example": "data: {\"type\": \"start\", \"question\": \"...\", \"mode\": \"replay\"}\n\ndata: {\"type\": \"step\", \"n\": 1, \"tool\": \"run_sql\", \"summary\": \"10 rows back, saved as q1\", \"body\": \"SELECT ...\"}\n\ndata: {\"type\": \"done\", \"result\": {\"answer\": \"...\", \"queries\": [...], \"charts\": [...]}}\n\n"
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/story": {
   "post": {
    "tags": [
     "Ask chat"
    ],
    "summary": "Ask the course chat (local)",
    "description": "Answers only from the brief, the build guide and the history (docs/), plus the passages that best match the\nquestion from the course's training guide and starting notebook and our notebook, cited as [1], [2]. Uses the\nAI service in `.env`. Errors come back as `{\"error\": \"...\"}`.",
    "operationId": "story_chat_api_story_post",
    "requestBody": {
     "description": "`question` (up to 500 characters) and optional `history`: the last few turns, oldest first.",
     "content": {
      "application/json": {
       "schema": {
        "type": "object"
       },
       "example": {
        "question": "How do I do Step 2?",
        "history": [
         {
          "role": "user",
          "content": "What does the brief ask for?"
         },
         {
          "role": "assistant",
          "content": "It asks for an agent that..."
         }
        ]
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "The answer, and the passages it cited.",
      "content": {
       "application/json": {
        "schema": {},
        "example": {
         "answer": "The training guide says the free plan allows about 8,000 tokens a minute [1]...",
         "sources": [
          {
           "n": 1,
           "source": "Course training guide",
           "title": "Free-tier limits",
           "url": ""
          }
         ],
         "model": "openai \u00b7 gpt-5.4-mini"
        }
       }
      }
     }
    }
   }
  },
  "/api/story/stream": {
   "post": {
    "tags": [
     "Ask chat"
    ],
    "summary": "Ask the course chat, streamed (local)",
    "description": "Same as `POST /api/story`, but the answer arrives word by word.",
    "operationId": "story_chat_stream_api_story_stream_post",
    "requestBody": {
     "description": "`question` (up to 500 characters) and optional `history`: the last few turns, oldest first.",
     "content": {
      "application/json": {
       "schema": {
        "type": "object"
       },
       "example": {
        "question": "How do I do Step 2?",
        "history": [
         {
          "role": "user",
          "content": "What does the brief ask for?"
         },
         {
          "role": "assistant",
          "content": "It asks for an agent that..."
         }
        ]
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Server-Sent Events: `delta` events with each piece of the answer as it's written, then `done` with the whole `answer`, its `sources` and the `model` (or `error`).",
      "content": {
       "text/event-stream": {
        "example": "data: {\"type\": \"delta\", \"text\": \"The training guide\"}\n\ndata: {\"type\": \"done\", \"answer\": \"...\", \"sources\": [], \"model\": \"openai \u00b7 gpt-5.4-mini\"}\n\n"
       }
      }
     }
    }
   }
  },
...
  "/api/sql": {
   "post": {
    "tags": [
     "Tools (no AI)"
    ],
    "summary": "Run one read-only SELECT",
    "description": "The `run_sql` tool: `columns`, `rows`, `row_count` and a `result_id` the agent can chart. Anything that\nchanges data is refused: try `DELETE FROM SalesOrderHeader` to see the `error`.",
    "operationId": "sql_api_sql_post",
    "requestBody": {
     "description": "`query`: one SELECT (or WITH \u2026 SELECT) on the `sales_lines` view or the raw tables.",
     "content": {
      "application/json": {
       "schema": {
        "type": "object"
       },
       "example": {
        "query": "SELECT channel, ROUND(SUM(revenue)) AS revenue FROM sales_lines GROUP BY channel"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   }
  },
...
  "/api/competitors/use": {
   "post": {
    "tags": [
     "Competitor prices"
    ],
    "summary": "Attach the mock list, your upload, or none",
    "description": "{\"source\": \"mock\" | \"uploaded\" | null}: which list your own questions can see.",
    "operationId": "competitors_use_api_competitors_use_post",
    "requestBody": {
     "description": "`source`: `mock`, `uploaded` or `null` (none).",
     "content": {
      "application/json": {
       "schema": {
        "type": "object"
       },
       "example": {
        "source": "mock"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 },
...
 "servers": [
  {
   "url": "http://localhost:8501",
   "description": "The demo app on your machine (./demo.sh app)"
  }
 ]
}